Added CreatedBy in Group 2/2. Added Posts in Group. Added Authorization. Added FluentValidators.

+++ SMS/Controllers/GroupsController.cs
@@ -2,20 +2,23 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using FluentValidation.Results;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SMS.Data;
 using SMS.Models;
+using SMS.Validators;
 
 namespace SMS.Controllers
 {
+    [Authorize]
     public class GroupsController : Controller
     {
         private readonly SMSContext _context;
         private readonly UserManager<SMSUser> _userManager;
-
 
         public GroupsController(SMSContext context, UserManager<SMSUser> userManager)
         {
@@ -38,13 +41,18 @@ namespace SMS.Controllers
 
             var @group = await _context.Groups
                 .Include(g => g.Subject)
+                .Include(g => g.Posts)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (@group == null)
             {
                 return NotFound();
             }
-
+            ViewBag.GroupId = id;
             return View(@group);
+        }
+        public IActionResult DetailsPost(int id)
+        {
+            return RedirectToAction("Details", "Posts", new { id = id });
         }
 
         public IActionResult Create()
@@ -58,6 +66,17 @@ namespace SMS.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Description,CreatedBy,SubjectId")] Group @group)
         {
+            var validator = new GroupValidator();
+            ValidationResult result = validator.Validate(group);
+
+            if (!result.IsValid)
+            {
+                foreach (var error in result.Errors)
+                {
+                    ModelState.AddModelError("", error.ErrorMessage);
+                }
+            }
+
             var userId = _userManager.GetUserId(HttpContext.User);
             SMSUser user = await _userManager.FindByIdAsync(userId);
 
@@ -65,7 +84,7 @@ namespace SMS.Controllers
             {
                 Name = @group.Name,
                 Description = @group.Description,
-                CreatedBy = user.FirstName,
+                CreatedBy = user.UserName,
                 SubjectId = @group.SubjectId
             };
 
@@ -99,33 +118,45 @@ namespace SMS.Controllers
                 return NotFound();
             }
 
-            ModelState.Remove("Subject");
-            ModelState.Remove("CreatedBy");
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    var existingGroup = await _context.Groups.FindAsync(id);
-                    group.CreatedBy = existingGroup.CreatedBy;
+            var validator = new GroupValidator();
+            ValidationResult result = validator.Validate(group);
 
-                    _context.Update(group);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
+            if (!result.IsValid)
+            {
+                foreach (var error in result.Errors)
                 {
-                    if (!GroupExists(group.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", error.ErrorMessage);
                 }
-                return RedirectToAction(nameof(Index));
             }
 
-            return View(group);
+            try
+            {
+                var groupToUpdate = await _context.Groups.FindAsync(id);
+                groupToUpdate.Name = group.Name;
+                groupToUpdate.Description = group.Description;
+                groupToUpdate.SubjectId = group.SubjectId;
+
+                _context.Update(groupToUpdate);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!GroupExists(group.Id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            return RedirectToAction(nameof(Index));
+            
+        }
+        public IActionResult EditPost(int id)
+        {
+            return RedirectToAction("Edit", "Posts", new { id = id });
         }
 
         public async Task<IActionResult> Delete(int? id)
@@ -150,18 +181,25 @@ namespace SMS.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Groups == null)
+            if (_context.Groups == null || _context.Posts == null)
             {
-                return Problem("Entity set 'ApplicationDbContext.Groups'  is null.");
+                return Problem("Entity set 'SMSContext.Groups' or 'SMSContext.Posts' is null.");
             }
             var @group = await _context.Groups.FindAsync(id);
             if (@group != null)
             {
+                var posts = _context.Posts.Where(p => p.GroupId == id);
+                _context.Posts.RemoveRange(posts);
                 _context.Groups.Remove(@group);
             }
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+        }
+
+        public IActionResult DeletePost(int id)
+        {
+            return RedirectToAction("Delete", "Posts", new { id = id });
         }
 
         private bool GroupExists(int id)
